traveltube/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
atomicwrites==1.4.0
attrs==22.1.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==40.0.1
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.14.0
django-debug-toolbar==2.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==12.0.1
flake8==3.9.2
flake8-docstrings==1.6.0
idna==2.8
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.6.1
mixer==7.1.2
more-itertools==9.0.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pyaes==1.6.1
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==3.0.9
Pyrogram==2.0.102
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.6
webcolors==1.11.1
whitenoise==2.0.6
